requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
dacktool==0.0.7
dbstream==0.1.21
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-allauth==0.58.2
django-appconf==1.0.6
django-autoslug==1.9.9
django-compressor==4.4
django-extensions==3.2.3
django-import-export==3.3.7
et-xmlfile==1.1.0
Faker==23.3.0
google-api-core==2.14.0
google-api-python-client==1.7.11
google-auth==2.23.4
google-auth-httplib2==0.0.3
google-auth-oauthlib==1.1.0
google-cloud-secret-manager==2.7.2
googleapis-common-protos==1.61.0
googleauthentication==0.0.17
grpc-google-iam-v1==0.12.7
grpcio==1.59.2
grpcio-status==1.59.2
httplib2==0.22.0
idna==3.4
import-export==0.3.1
install==1.3.5
libcst==1.1.0
MarkupPy==1.14
mypy-extensions==1.0.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
paypalrestsdk==1.13.3
Pillow==10.1.0
proto-plus==1.22.3
protobuf==4.25.0
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydotplus==2.0.2
PyJWT==2.8.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
PyYAML==6.0.1
rcssmin==1.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==3.0.1
urllib3==2.0.7
xlrd==2.0.1
xlwt==1.3.0
